+++ mink/sparv/models.py
@@ -248,8 +248,10 @@ statuses_response_examples = [
                 "info": "Job was completed successfully",
                 "resource": {
                     "id": "mink-ezodmp4wxm",
+                    "public_id": "mink-ezodmp4wxm",
                     "name": {"swe": "txt-korpus", "eng": "txt-korpus"},
                     "type": "corpus",
+                    "custom_config": False,
                     "source_files": [
                         {
                             "name": "text1.txt",
@@ -272,43 +274,9 @@ statuses_response_examples = [
             {
                 "job_status": "done",
                 "info": "Job was completed successfully",
-                "resource": {
-                    "id": "mink-dxh6e6wtff",
-                    "name": {"swe": "Annes och Martins testkorpus", "eng": ""},
-                    "type": "corpus",
-                    "source_files": [
-                        {
-                            "name": "dokument2.xml",
-                            "type": "text/xml",
-                            "last_modified": "2022-12-22T11:25:25+01:00",
-                            "size": 115,
-                            "path": "dokument2.xml",
-                        },
-                        {
-                            "name": "dokument3.xml",
-                            "type": "text/xml",
-                            "last_modified": "2023-06-13T13:26:44+02:00",
-                            "size": 41,
-                            "path": "dokument3.xml",
-                        },
-                        {
-                            "name": "dokument4.xml",
-                            "type": "text/xml",
-                            "last_modified": "2023-06-13T13:26:44+02:00",
-                            "size": 461,
-                            "path": "dokument4.xml",
-                        },
-                        {
-                            "name": "dokument1.xml",
-                            "type": "text/xml",
-                            "last_modified": "2023-06-13T13:26:49+02:00",
-                            "size": 1394,
-                            "path": "dokument1.xml",
-                        },
-                    ],
-                },
+                "resource": models.resource_model_example,
                 "job": job_model_examples[1],
-            },
+            }
         ],
     }
 ]